css/selectors/simple-selectors/pseudo-classes/css3-selectors-lang-005.html

<!DOCTYPE html>
<html  lang="en" >
<head>
<meta charset="utf-8"/>
<title>CSS3 Selectors: :lang(es), lang="es-MX"</title>
<link rel='author' title='Richard Ishida' href='mailto:ishida@w3.org'>
<link rel='help' href='http://www.w3.org/TR/selectors/Overview.html#lang-pseudo'>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<link href="../../../resources/testharness.css" rel="stylesheet" type="text/css">
<meta name='flags' content=' :lang'>
<meta name="assert" content="A :lang value will match a lang attribute value when the latter contains additional subtags.">
<style type='text/css'>
.test div { width: 50px; }
 		#colonlangcontroltest { color: red; font-weight: bold; width: 400px; }
		#colonlangcontroltest:lang(xx) { display:none; } 
#box:lang(es) { width: 100px; }
</style>
</head>
<body>



<div class="test"><div id="box" lang="es-MX">&#xA0;</div></div>
<p lang='xx' id='colonlangcontroltest'>This test failed because it relies on :lang for results, but :lang is not supported by this browser.


<!--Notes:
This tests a detail related to :lang support. If :lang is not supported, a message will appear and the test will fail.

-->
<script>
test(function() {
assert_equals(document.getElementById('colonlangcontroltest').offsetWidth, 0)
assert_equals(document.getElementById('box').offsetWidth, 100);
}, " ");
</script>

<div id='log'></div>

</body>
</html>
